requirements.txt
streamlit>=1.28.0
pandas>=1.4.1
numpy>=1.22.0
matplotlib>=3.5.1
seaborn>=0.11.2
scikit-learn>=1.1.2
gqlalchemy>=1.3.1
pyarrow>=7.0.0
python-dotenv>=0.19.2
PyYAML>=6.0